+++ tests/testthat/helper-data.R
@@ -6,6 +6,7 @@ lambda <- 0.9 # 0.9 # 1 / 0.9
 dates <- rev(seq(Sys.Date(), length.out = n_obs, by = "-1 day"))
 
 # test arguments
+test_p <- c(0, 0.25, 0.5, 0.75, 1)
 test_width <- c(1, 5, 10, 15)
 test_intercept <- c(TRUE, FALSE)
 test_center <- c(TRUE, FALSE)